set_bb_env: use ncurse based tool to query MACHINE and DISTRO
Uses ncurse (whiptail or dialog tool) to query the user with a list of
MACHINE and DISTRO to pick. We build the list of valid MACHINE and
DISTRO for now by looking at meta-qti-bsp, but that could change.

MACHINE and DISTRO are listed along with the name of their layer. The
user can simply select the one needed for the build, eventually we
should set BUILDDIR to be build-DISTRO or build-DISTRO-MACHINE.

Then the values for MACHINE and DISTRO are put in conf/auto.conf,
which is generated and overriden.

Then the user can simply use bitbake commands as usual.

It is possible to set MACHINE and DISTRO before running set_bb_env, so
that it can work non interactively as well, eg.

MACHINE=foo DISTRO=bar source build/conf/set_bb_env.sh

In that case ncurse is not used, but conf/auto.conf is still
generated.

MACHINE and DISTRO are left in the environment for now, so that
running set_bb_env.sh again won't ask user for the choice, but
unset_bb_env would clean that up.

Change-Id: Ieacb3ad6099e1d72b053bed60b1ac25379ebdaac
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>