set_bb_env: allows custom BUILDDIR
By default BUILDDIR is set to build-DISTRO, however now that
SSTATE_CACHE and DL_DIR are outside of the build folder, it can be
convenient to create many build folders to run different builds or
configs.

With this change, we can run:
source <path>/set_env_bb.sh build-foobar

and it will create a build environment in build-foobar, we can have as
many build folders as needed, and each build folder has its own config
files to allow various configs to be built/tested in parallel.

If BUILDDIR is set on the command line, and hapens to be a valid build
environment, then all configuration is skipped and we simply use the
folder as is.

Change-Id: I27e4187c86fc1ef6594b0624a6940caafa0dcae1
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>