local.conf: fix BB_HASHBASE_WHITELIST
BB_HASHBASE_WHITELIST is redefined here, which means we are not using
the default value from meta/conf/bitbake.conf. In Sumo this change was
done:
https://git.openembedded.org/openembedded-core/commit/meta/conf/bitbake.conf?h=sumo&id=8547f1e29104b75299f1056524da4a058a029940

It was never brought to our copy of BB_HASHBASE_WHITELIST , and is
still missing as of today.

There is a comment here says "Because appending to it was not
working", however I have tested using _append for
BB_HASHBASE_WHITELIST and it is working as expected.

Change-Id: I3629ed788ca65c7bc864cc63f1a86c3a24805b7a
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>