qti-conf: add support for BB_SIGNATURE_LOCAL_DIRS_EXCLUDE
This patch replaces the initial attempt to avoid taking .git directory
content into account when using the local fetcher. The initial
implementation was trying to detect if the 'source' was a git repo and
in which case it was using git HEAD & diff to create a pseudo
checksum. This approach was rejected upstream, and instead a new
implemtation is available. This patch is bringing this new
implementation into our build.

Instead of using git HEAD & diff within the local fetcher, we now
ignore and VCS folder (.git, .svn, ..) when traversing the tree to
compute file checksums, that ensures that the content of .git is not
taken into account.

A few additional changes are included in this patch:
* use a 'series' file to control which patch are applied in
set_bb_env, to avoid applying any local patches that might be lying in
the patches folder.
* use -p1 instead of -p0 for patches, since it's more common when
dealing with patches extracted from mailing lists and/or git.
* set BB_SIGNATURE_LOCAL_DIRS_EXCLUDE in local.conf. This new
implementation relies on this variable to indicate what are the
folders to ignore, it is not set by default in bitbake, so we need to
set it explicitely.

Change-Id: I713f199035d1f424a188f190eebb97a27b9efc80
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>