Disable sdk build for quin-gvm-gen4/quin-gvm-gen4-sdk
There is an known UID confilct error in sdk build when qti-display
enabled and qti-audio disabled.

    groupadd: GID '4054' already exists
    ERROR: binder: groupadd command did not succeed.

The root cause is binder is included in sdk as display-noship-linux DEPENDS
to binder. But binder is not installed into image rootfs, and GID 4054
is taken by weston.

Disable sdk build for quin-gvm-gen4/quin-gvm-gen4-sdk temporarily.

Change-Id: I2e0d7e5a37968b3e14eb9a9f3a7117c2c6a8d439
Signed-off-by: xb <quic_xubaoj@quicinc.com>